requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
psycopg2==2.9.9
gunicorn==23.0.0
six==1.17.0
celery==5.5.3
requests==2.32.4
pymemcache==4.0.0
celery-amqp-backend==1.2.0
pyjwt[crypto]==2.8.0
